pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "xaihack"
version = "0.1.0"
description = "Voice agents for dialectical reasoning"
requires-python = ">=3.11"
dependencies = [
    # for voiceagents only
    "livekit==1.0.20",
    "livekit-agents==1.3.6",
    "livekit-plugins-deepgram==1.3.6",
    "livekit-plugins-openai==1.3.6",
    "livekit-plugins-elevenlabs==1.3.6",
    "livekit-plugins-silero==1.3.6",
    "livekit-plugins-turn-detector==1.3.6",
    "livekit-plugins-noise-cancellation==0.2.5",
    "python-dotenv",
    "langchain-xai",
    "pydantic>=2.0",
    "diskcache>=5.6.3",
    "tenacity>=9.1.2",
    "websockets>=12.0",
]

[project.scripts]
voiceagents = "voiceagents.app:main"

[tool.hatch.build.targets.wheel]
packages = ["voiceagents"]

